+++ src/traits/mod.rs
@@ -1,12 +1,9 @@
 mod contextual;
 mod conversion;
 mod into_as;
-mod owenance;
 
 pub use contextual::ErrorWith;
 pub use conversion::{
     ConvStructError, ErrorConv, ErrorWrapAs, ToStructError, WrapStructErrorAs,
 };
 pub use into_as::{raw_source, IntoAs, RawSource, RawStdError};
-#[allow(deprecated)]
-pub use owenance::{ErrorOwe, ErrorOweBase, ErrorOweSource, ErrorOweSourceBase};